+++ src/main/java/dev/fenix/application/resume/repository/TechnologieRepository.java
@@ -3,8 +3,8 @@ package dev.fenix.application.resume.repository;
 import dev.fenix.application.resume.model.Technologie;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TechnologieRepository extends JpaRepository<Technologie,Integer> {
-    Technologie getIconById(Integer id);
+public interface TechnologieRepository extends JpaRepository<Technologie,Long> {
+    Technologie getIconById(Long id);
     Technologie getIconByName(String title);
     Technologie save(Technologie b);
     void delete(Technologie b);
